relocate encoding
