+++ crates/rust-eigenda-v2-client/Cargo.toml
@@ -16,7 +16,8 @@ exclude = [
 
 [dependencies]
 rust-eigenda-signers = { version = "0.1.6" }
-rust-eigenda-v2-common = "0.1.2"
+# rust-eigenda-v2-common = "0.1.2"
+rust-eigenda-v2-common = { path = "../rust-eigenda-v2-common"}
 
 rand = { workspace = true }
 bytes = { workspace = true }
