+++ crates/rust-eigenda-v2-common/Cargo.toml
@@ -18,6 +18,9 @@ ark-serialize = { workspace = true }
 rust-kzg-bn254-primitives = { workspace = true }
 bincode = { workspace = true }
 thiserror = { workspace = true }
+alloy ={ workspace = true }
+ethabi = { workspace = true }
+tiny-keccak = { workspace = true }
 
 [dev-dependencies]
 proptest = { workspace = true }
